lots of shuffling, renaming, and access changes
