+++ Content.Client/NetworkConfigurator/DeviceListSystem.cs
@@ -6,48 +6,16 @@ namespace Content.Client.NetworkConfigurator;
 
 public sealed class DeviceListSystem : SharedDeviceListSystem
 {
-    [Dependency] private readonly IOverlayManager _overlay = default!;
-
     /// <summary>
     /// Toggles the given device lists connection visualisation on and off.
     /// TODO: Implement an overlay that draws a line between the given entity and the entities in the device list
     /// </summary>
-    public void ToggleVisualization(EntityUid uid, bool toggle, NetworkConfiguratorComponent? component = null)
+    public IEnumerable<EntityUid> GetAllDevices(EntityUid uid, DeviceListComponent? component = null)
     {
-        if (!Resolve(uid, ref component) || component.ActiveDeviceList == null)
-            return;
-
-        if (!toggle)
+        if (!Resolve(uid, ref component))
         {
-            RemComp<NetworkConfiguratorActiveLinkOverlayComponent>(component.ActiveDeviceList.Value);
-            if (!EntityQuery<NetworkConfiguratorActiveLinkOverlayComponent>().Any())
-            {
-                _overlay.RemoveOverlay<NetworkConfiguratorLinkOverlay>();
-            }
-
-            return;
+            return new EntityUid[] { };
         }
-
-        if (!_overlay.HasOverlay<NetworkConfiguratorLinkOverlay>())
-        {
-            _overlay.AddOverlay(new NetworkConfiguratorLinkOverlay());
-        }
-
-        EnsureComp<NetworkConfiguratorActiveLinkOverlayComponent>(component.ActiveDeviceList.Value);
-    }
-
-    public void ClearAllOverlays()
-    {
-        if (!_overlay.HasOverlay<NetworkConfiguratorLinkOverlay>())
-        {
-            return;
-        }
-
-        foreach (var tracker in EntityQuery<NetworkConfiguratorActiveLinkOverlayComponent>())
-        {
-            RemCompDeferred<NetworkConfiguratorActiveLinkOverlayComponent>(tracker.Owner);
-        }
-
-        _overlay.RemoveOverlay<NetworkConfiguratorLinkOverlay>();
+        return component.Devices;
     }
 }
