+++ Content.Client/NetworkConfigurator/NetworkConfiguratorBoundUserInterface.cs
@@ -9,7 +9,7 @@ public sealed class NetworkConfiguratorBoundUserInterface : BoundUserInterface
     private NetworkConfiguratorListMenu? _listMenu;
     private NetworkConfiguratorConfigurationMenu? _configurationMenu;
 
-    private DeviceListSystem? _deviceList;
+    private NetworkConfiguratorSystem? _netConfig;
 
     public NetworkConfiguratorBoundUserInterface(ClientUserInterfaceComponent owner, Enum uiKey) : base(owner, uiKey)
     {
@@ -24,7 +24,7 @@ public sealed class NetworkConfiguratorBoundUserInterface : BoundUserInterface
     {
         base.Open();
 
-        _deviceList = IoCManager.Resolve<IEntityManager>().System<DeviceListSystem>();
+        _netConfig = IoCManager.Resolve<IEntityManager>().System<NetworkConfiguratorSystem>();
 
         switch (UiKey)
         {
@@ -42,7 +42,7 @@ public sealed class NetworkConfiguratorBoundUserInterface : BoundUserInterface
                 //_configurationMenu.Edit.OnPressed += _ => OnConfigButtonPressed(NetworkConfiguratorButtonKey.Edit);
                 _configurationMenu.Clear.OnPressed += _ => OnConfigButtonPressed(NetworkConfiguratorButtonKey.Clear);
                 _configurationMenu.Copy.OnPressed += _ => OnConfigButtonPressed(NetworkConfiguratorButtonKey.Copy);
-                _configurationMenu.Show.OnPressed += args => _deviceList.ToggleVisualization(Owner.Owner, args.Button.Pressed);
+                _configurationMenu.Show.OnPressed += args => _netConfig.ToggleVisualization(Owner.Owner, args.Button.Pressed);
                 _configurationMenu.OpenCentered();
                 break;
         }
