--- Content.Client/NetworkConfigurator/NetworkConfiguratorComponent.cs
@@ -1,11 +1,0 @@
-using Content.Shared.DeviceNetwork;
-
-namespace Content.Client.NetworkConfigurator;
-
-/// <summary>
-/// This is used for...
-/// </summary>
-[RegisterComponent]
-public sealed class NetworkConfiguratorComponent : SharedNetworkConfiguratorComponent
-{
-}
