+++ Content.Shared/DeviceNetwork/SharedNetworkConfiguratorSystem.cs
@@ -1,0 +1,31 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.DeviceNetwork;
+
+public abstract class SharedNetworkConfiguratorSystem : EntitySystem
+{
+    public override void Initialize()
+    {
+        base.Initialize();
+
+        SubscribeLocalEvent<NetworkConfiguratorComponent, ComponentGetState>(GetNetworkConfiguratorState);
+        SubscribeLocalEvent<NetworkConfiguratorComponent, ComponentHandleState>(HandleNetworkConfiguratorState);
+    }
+
+    private void GetNetworkConfiguratorState(EntityUid uid, NetworkConfiguratorComponent comp,
+        ref ComponentGetState args)
+    {
+        args.State = new NetworkConfiguratorComponentState(comp.ActiveDeviceList);
+    }
+
+    private void HandleNetworkConfiguratorState(EntityUid uid, NetworkConfiguratorComponent comp,
+        ref ComponentHandleState args)
+    {
+        if (args.Current is not NetworkConfiguratorComponentState state)
+        {
+            return;
+        }
+
+        comp.ActiveDeviceList = state.ActiveDeviceList;
+    }
+}
